+++ packages/google_maps_flutter/google_maps_flutter/CHANGELOG.md
@@ -1,6 +1,14 @@
 ## 2.3.0
 
 * Adds support for heatmap layers.
+
+## 2.2.5
+
+* Updates iOS minimum version in README.
+
+## 2.2.4
+
+* Updates links for the merge of flutter/plugins into flutter/packages.
 * Updates minimum Flutter version to 3.0.
 
 ## 2.2.3