+++ packages/google_maps_flutter/google_maps_flutter_ios/CHANGELOG.md
@@ -1,6 +1,14 @@
-## 2.2.0
+## 2.3.0
 
 * Adds support for heatmap layers.
+
+## 2.2.0
+
+* Updates minimum Flutter version to 3.3 and iOS 11.
+
+## 2.1.14
+
+* Updates links for the merge of flutter/plugins into flutter/packages.
 * Updates minimum Flutter version to 3.0.
 
 ## 2.1.13