+++ packages/google_maps_flutter/google_maps_flutter_web/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 0.4.1
 
 * Adds support for heatmap layers.
+
+## 0.4.0+6
+
+* Updates links for the merge of flutter/plugins into flutter/packages.
 * Updates minimum Flutter version to 3.0.
 
 ## 0.4.0+5