+++ packages/shared_preferences/shared_preferences_android/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## NEXT
 
+* Updates compileSdkVersion to 33.
+
+## 2.0.16
+
+* Updates links for the merge of flutter/plugins into flutter/packages.
 * Updates minimum Flutter version to 3.0.
 
 ## 2.0.15