+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,5 +1,10 @@
-## NEXT
+## 2.5.3
 
+* Updates iOS minimum version in README.
+
+## 2.5.2
+
+* Updates links for the merge of flutter/plugins into flutter/packages.
 * Updates minimum Flutter version to 3.0.
 
 ## 2.5.1