+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.3.2
+
+* Updates compileSdkVersion to 33.
+* Bumps androidx.webkit:webkit from 1.5.0 to 1.6.0.
+
+## 3.3.1
+
+* Updates links for the merge of flutter/plugins into flutter/packages.
+
 ## 3.3.0
 
 * Adds support to access native `WebView`.