+++ packages/webview_flutter/webview_flutter_wkwebview/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.2.0
+
+* Updates minimum Flutter version to 3.3 and iOS 11.
+
+## 3.1.1
+
+* Updates links for the merge of flutter/plugins into flutter/packages.
+
 ## 3.1.0
 
 * Adds support to access native `WKWebView`.